Server testing partially implemented
Refactored the server into a class, so I can now set up and tear down a
server in my test module.

The point of entry for the app is now app.js.

Created server_spec.js which runs a couple of tests to make sure socket events
are being responded to accordingly.