Corrected error in server handling disconnections
The test for the 'user left' event, which fires when a user disconnects
was failing. The reason was in the server handler for the disconnect
event, there was a check to see if the addedUser was set, which wouldn't
be set in the test environment.

The whole event emmit function was inside the if-statement for this check
so it wasn't being executed by the test events. The only useful thing that
happens after that check which needs to be user-related, is decrementing
the numUsers variable, so this was left inside the check.